requirements.txt
asgiref==3.6.0
Django==3.2.19
django-cors-headers==4.3.1
djangorestframework==3.15.1
python-dotenv==1.0.1
sqlparse==0.4.2
